Use buffer_ptr inside predicate skips
While I can't reproduce this bug in the test suite, I suspect that in some cases the value decoding will run over the end of the buffer.

This is because the buffer is not NUL terminated, and the code was originally expecting to perform this in-situ on the original buffer, which was both NUL-terminated and within known delimiters.